backend/requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.4,<4.5
psycopg[binary]>=3.2,<3.3
python-decouple>=3.8,<3.9